json fix 2
The issue was that the System.Text.Json serializer doesn't properly serialize ValueTuples by default. When a participant was added to a signup, the ParticipantInfo was being created as a ValueTuple (ulong Id, string Username), but when serialized to JSON, it was being written as an empty object {} instead of the expected properties.
By replacing ValueTuples with proper classes that have explicit property names, we ensure that the serializer can correctly serialize and deserialize the participant information. This should fix both the issue with participants not being stored in the JSON file and not showing up in the embed.
You should now be able to deploy these changes to your server and see the participants correctly displayed in the signup messages and stored in the signups.json file.